Abort with a 404 code if the member isn't part of a team when you try to remove them@

+++ app/Http/Controllers/Teams/TeamMembersController.php
@@ -28,6 +28,11 @@ class TeamMembersController extends Controller
 
     public function remove(Team $team, User $member): RedirectResponse
     {
+        // If member isn't in team
+        if (!$team->members->contains($member)){
+            abort(404);
+        }
+
         // Remove the member from the team
         $team->members()->detach($member);
